refactor: rename PaneInfo.vue
