+++ src/components/PanelRight/PaneNodeInfo.vue
